Adiciando swagger
* Adicionado swagger config;
* Renomeado nome do projeto;
* Corrigido problema ao salvar data no banco de dados;
