+++ backend/bin/main/.gitignore
@@ -1,2 +1,1 @@
-/application.properties
 /br/
