+++ backend/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'backend'
+rootProject.name = 'reembolsocoders'
