--- backend/src/main/java/br/com/academiadev/projetocoders/backend/converter/Converter.java
@@ -1,9 +1,0 @@
-package br.com.academiadev.projetocoders.backend.converter;
-
-public interface Converter<T, D> {
-	
-	public D toDTO(T entity);
-
-	public T toEntity(D dto);
-
-}
