--- backend/src/main/java/br/com/academiadev/projetocoders/backend/dto/EmpresaDTO.java
@@ -1,45 +1,0 @@
-package br.com.academiadev.projetocoders.backend.dto;
-
-public class EmpresaDTO {
-
-	private String nome;
-
-	private String email;
-
-	private Long codigo;
-
-	private String senha;
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public Long getCodigo() {
-		return codigo;
-	}
-
-	public void setCodigo(Long codigo) {
-		this.codigo = codigo;
-	}
-
-	public String getSenha() {
-		return senha;
-	}
-
-	public void setSenha(String senha) {
-		this.senha = senha;
-	}
-
-}
