--- backend/src/main/java/br/com/academiadev/projetocoders/backend/dto/FuncionarioDTO.java
@@ -1,57 +1,0 @@
-package br.com.academiadev.projetocoders.backend.dto;
-
-import java.time.LocalDate;
-
-public class FuncionarioDTO {
-
-	private String nome;
-
-	private String email;
-
-	private EmpresaDTO empresa;
-
-	private String senha;
-
-	private LocalDate dataCadastro;
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public EmpresaDTO getEmpresa() {
-		return empresa;
-	}
-
-	public void setEmpresa(EmpresaDTO empresa) {
-		this.empresa = empresa;
-	}
-
-	public String getSenha() {
-		return senha;
-	}
-
-	public void setSenha(String senha) {
-		this.senha = senha;
-	}
-
-	public LocalDate getDataCadastro() {
-		return dataCadastro;
-	}
-
-	public void setDataCadastro(LocalDate dataCadastro) {
-		this.dataCadastro = dataCadastro;
-	}
-
-}
