--- backend/src/main/java/br/com/academiadev/projetocoders/backend/dto/ReembolsoDTO.java
@@ -1,85 +1,0 @@
-package br.com.academiadev.projetocoders.backend.dto;
-
-public class ReembolsoDTO {
-
-	private Long id;
-
-	private String descricao;
-
-	private String categoria;
-
-	private String data;
-
-	private StatusDTO status;
-
-	private FuncionarioDTO funcionario;
-
-	private String arquivoPath;
-
-	private String valor;
-
-	public String getValor() {
-		return valor;
-	}
-
-	public void setValor(String valor) {
-		this.valor = valor;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getDescricao() {
-		return descricao;
-	}
-
-	public void setDescricao(String descricao) {
-		this.descricao = descricao;
-	}
-
-	public String getCategoria() {
-		return categoria;
-	}
-
-	public void setCategoria(String categoria) {
-		this.categoria = categoria;
-	}
-
-	public String getData() {
-		return data;
-	}
-
-	public void setData(String data) {
-		this.data = data;
-	}
-
-	public StatusDTO getStatus() {
-		return status;
-	}
-
-	public void setStatus(StatusDTO status) {
-		this.status = status;
-	}
-
-	public FuncionarioDTO getFuncionario() {
-		return funcionario;
-	}
-
-	public void setFuncionario(FuncionarioDTO funcionario) {
-		this.funcionario = funcionario;
-	}
-
-	public String getArquivoPath() {
-		return arquivoPath;
-	}
-
-	public void setArquivoPath(String arquivoPath) {
-		this.arquivoPath = arquivoPath;
-	}
-
-}
