--- backend/src/main/java/br/com/academiadev/projetocoders/backend/dto/StatusDTO.java
@@ -1,15 +1,0 @@
-package br.com.academiadev.projetocoders.backend.dto;
-
-public class StatusDTO {
-
-	private String status;
-
-	public String getStatus() {
-		return status;
-	}
-
-	public void setStatus(String status) {
-		this.status = status;
-	}
-
-}
