--- backend/src/main/java/br/com/academiadev/projetocoders/backend/exception/EmpresaNaoEncontradaException.java
@@ -1,5 +1,0 @@
-package br.com.academiadev.projetocoders.backend.exception;
-
-public class EmpresaNaoEncontradaException extends Exception {
-
-}
