--- backend/src/main/java/br/com/academiadev/projetocoders/backend/model/Empresa.java
@@ -1,81 +1,0 @@
-package br.com.academiadev.projetocoders.backend.model;
-
-import java.util.List;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.validation.constraints.NotNull;
-
-@Table(name = "empresa")
-@Entity
-public class Empresa {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	@Column
-	private Long id;
-
-	@Column
-	@NotNull
-	private String nome;
-
-	@Column
-	@NotNull
-	private String email;
-
-	@Column
-	@NotNull
-	private Long codigo;
-
-	@Column
-	@NotNull
-	private String senha;
-
-	@OneToMany(mappedBy = "empresa")
-	private List<Funcionario> funcionarios;
-
-	public Long getCodigo() {
-		return codigo;
-	}
-
-	public void setCodigo(Long codigo) {
-		this.codigo = codigo;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getSenha() {
-		return senha;
-	}
-
-	public void setSenha(String senha) {
-		this.senha = senha;
-	}
-
-	public List<Funcionario> getFuncionarios() {
-		return funcionarios;
-	}
-
-	public void setFuncionarios(List<Funcionario> funcionarios) {
-		this.funcionarios = funcionarios;
-	}
-
-	public String getNome() {
-		return nome;
-	}
-
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-}
