--- backend/src/main/java/br/com/academiadev/projetocoders/backend/repository/EmpresaRepository.java
@@ -1,13 +1,0 @@
-package br.com.academiadev.projetocoders.backend.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import br.com.academiadev.projetocoders.backend.model.Empresa;
-
-@Repository
-public interface EmpresaRepository extends CrudRepository<Empresa, Long>{
-	
-	public Empresa findByCodigo(Long codigo);
-
-}
