--- backend/src/main/java/br/com/academiadev/projetocoders/backend/repository/FuncionarioRepository.java
@@ -1,13 +1,0 @@
-package br.com.academiadev.projetocoders.backend.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import br.com.academiadev.projetocoders.backend.model.Funcionario;
-
-@Repository
-public interface FuncionarioRepository extends CrudRepository<Funcionario, Long>{
-	
-	public Funcionario findByNome(String nome);
-	
-}
