+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/ReembolsoApplication.java
@@ -1,0 +1,12 @@
+package br.com.academiadev.projetocoders.reembolsocoders;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReembolsoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReembolsoApplication.class, args);
+	}
+}
