+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/config/SwaggerConfig.java
@@ -1,0 +1,30 @@
+package br.com.academiadev.projetocoders.reembolsocoders.config;
+
+import java.util.ArrayList;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+public class SwaggerConfig {
+	
+	@Bean
+	public Docket apis() {
+		return new Docket(DocumentationType.SWAGGER_2).select()
+				.apis(RequestHandlerSelectors.basePackage("br.com.academiadev.projetocoders.reembolsocoders"))
+				.paths(PathSelectors.any()).build().apiInfo(apiInfo());
+	}
+	
+	private ApiInfo apiInfo() {
+        return new ApiInfo("titulo", "descricao", "versão", "termos", null, "Licence", "licenseUrl", new ArrayList<>());
+    }
+
+}
