+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/controller/ReembolsoController.java
@@ -1,0 +1,23 @@
+package br.com.academiadev.projetocoders.reembolsocoders.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import br.com.academiadev.projetocoders.reembolsocoders.dto.ReembolsoDTO;
+import br.com.academiadev.projetocoders.reembolsocoders.service.ReembolsoService;
+
+@RestController
+public class ReembolsoController {
+
+	@Autowired
+	private ReembolsoService reembolsoService;
+
+	@PostMapping("/cadastrarReembolso")
+	public void cadastrarReembolso(@RequestBody ReembolsoDTO reembolsoDTO, @RequestParam String nomeFuncionario) {
+		reembolsoService.Cadastrar(reembolsoDTO, nomeFuncionario);
+	}
+
+}
