+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/converter/Converter.java
@@ -1,0 +1,9 @@
+package br.com.academiadev.projetocoders.reembolsocoders.converter;
+
+public interface Converter<T, D> {
+	
+	public D toDTO(T entity);
+
+	public T toEntity(D dto);
+
+}
