+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/dto/StatusDTO.java
@@ -1,0 +1,15 @@
+package br.com.academiadev.projetocoders.reembolsocoders.dto;
+
+public class StatusDTO {
+
+	private String status;
+
+	public String getStatus() {
+		return status;
+	}
+
+	public void setStatus(String status) {
+		this.status = status;
+	}
+
+}
