+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/exception/EmpresaNaoEncontradaException.java
@@ -1,0 +1,5 @@
+package br.com.academiadev.projetocoders.reembolsocoders.exception;
+
+public class EmpresaNaoEncontradaException extends Exception {
+
+}
