+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/model/Funcionario.java
@@ -1,0 +1,111 @@
+package br.com.academiadev.projetocoders.reembolsocoders.model;
+
+import java.time.LocalDate;
+import java.util.List;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.validation.constraints.NotNull;
+
+@Table(name = "funcionario")
+@Entity
+public class Funcionario {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	@Column
+	private Long id;
+
+	@Column
+	private Integer cpf;
+
+	@Column
+	@NotNull
+	private String nome;
+
+	@Column
+	private String email;
+
+	@ManyToOne
+	private Empresa empresa;
+
+	@Column
+	@NotNull
+	private String senha;
+
+	@OneToMany(mappedBy = "funcionario")
+	private List<Reembolso> reembolsos;
+
+	@Column
+	private LocalDate dataCadastro;
+
+	public Empresa getEmpresa() {
+		return empresa;
+	}
+
+	public void setEmpresa(Empresa empresa) {
+		this.empresa = empresa;
+	}
+
+	public List<Reembolso> getReembolsos() {
+		return reembolsos;
+	}
+
+	public void setReembolsos(List<Reembolso> reembolsos) {
+		this.reembolsos = reembolsos;
+	}
+
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
+	public Integer getCpf() {
+		return cpf;
+	}
+
+	public void setCpf(Integer cpf) {
+		this.cpf = cpf;
+	}
+
+	public String getNome() {
+		return nome;
+	}
+
+	public void setNome(String nome) {
+		this.nome = nome;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	public String getSenha() {
+		return senha;
+	}
+
+	public void setSenha(String senha) {
+		this.senha = senha;
+	}
+
+	public LocalDate getDataCadastro() {
+		return dataCadastro;
+	}
+
+	public void setDataCadastro(LocalDate dataCadastro) {
+		this.dataCadastro = dataCadastro;
+	}
+}
