+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/repository/FuncionarioRepository.java
@@ -1,0 +1,13 @@
+package br.com.academiadev.projetocoders.reembolsocoders.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import br.com.academiadev.projetocoders.reembolsocoders.model.Funcionario;
+
+@Repository
+public interface FuncionarioRepository extends CrudRepository<Funcionario, Long>{
+	
+	public Funcionario findByNome(String nome);
+	
+}
