+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/repository/ReembolsoRepository.java
@@ -1,0 +1,11 @@
+package br.com.academiadev.projetocoders.reembolsocoders.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import br.com.academiadev.projetocoders.reembolsocoders.model.Reembolso;
+
+@Repository
+public interface ReembolsoRepository extends CrudRepository<Reembolso, Long> {
+
+}
