+++ backend/src/main/java/br/com/academiadev/projetocoders/reembolsocoders/service/EmpresaService.java
@@ -1,0 +1,34 @@
+package br.com.academiadev.projetocoders.reembolsocoders.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import br.com.academiadev.projetocoders.reembolsocoders.converter.EmpresaConverter;
+import br.com.academiadev.projetocoders.reembolsocoders.dto.EmpresaDTO;
+import br.com.academiadev.projetocoders.reembolsocoders.model.Empresa;
+import br.com.academiadev.projetocoders.reembolsocoders.repository.EmpresaRepository;
+
+@Service
+public class EmpresaService {
+
+	@Autowired
+	private EmpresaRepository empresaRepository;
+	
+	@Autowired
+	private EmpresaConverter empresaConverter;
+	
+	public void Cadastrar(EmpresaDTO empresaDTO) {
+		Empresa empresa = empresaConverter.toEntity(empresaDTO);		
+		empresaRepository.save(empresa);
+	}
+	
+	public EmpresaDTO CriarEmpresaDTO() {
+		EmpresaDTO empresaDTO = new EmpresaDTO();
+		empresaDTO.setCodigo(1010l);
+		empresaDTO.setEmail("empresa@gmail.com");
+		empresaDTO.setNome("Empresa");
+		empresaDTO.setSenha("123");
+		
+		return empresaDTO;
+	}
+}
